Fix opauth parse uri function
The parseUri function takes the URI of the request and extracts the
paramters. The request string starts immediately after the web path.

The current implementation does not consider, that the URI can start
with a port definition. In this case the extracted request is not
correct.

By searching the web path and let the paramter string start after that,
the URI is parsed correctly.